+++ packages/vega/src/VegaChartToolResult.tsx
@@ -1,27 +1,13 @@
 import {useStoreWithAi} from '@sqlrooms/ai';
-import {
-  Button,
-  cn,
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-  useDisclosure,
-} from '@sqlrooms/ui';
-import {Check, Code2, Database, EditIcon} from 'lucide-react';
+import {cn} from '@sqlrooms/ui';
 import {useCallback, useState} from 'react';
 import {EmbedOptions, VisualizationSpec} from 'vega-embed';
 import {VegaChartContainer} from './editor/VegaChartContainer';
 import {VegaChartDisplay} from './editor/VegaChartDisplay';
-import {useVegaEditorContext} from './editor/VegaEditorContext';
-import {VegaSpecEditorPanel} from './editor/VegaSpecEditorPanel';
-import {VegaSqlEditorPanel} from './editor/VegaSqlEditorPanel';
 import {EditorMode} from './editor/types';
-import {VegaLiteArrowChart} from './VegaLiteArrowChart';
+import {VegaEditAction} from './VegaEditAction';
 import {VegaExportAction} from './VegaExportAction';
+import {VegaLiteArrowChart} from './VegaLiteArrowChart';
 
 export type VegaChartToolResultProps = {
   className?: string;
@@ -46,42 +32,6 @@ export type VegaChartToolResultProps = {
 };
 
 /**
- * Inline editor actions component (Discard + Apply)
- */
-function EditorActions({onClose}: {onClose: () => void}) {
-  const {actions, canApply, hasChanges} = useVegaEditorContext();
-
-  return (
-    <div className="flex items-center gap-1">
-      <Button
-        variant="ghost"
-        size="xs"
-        onClick={() => {
-          actions.cancelChanges();
-        }}
-        disabled={!hasChanges}
-        title="Discard changes"
-      >
-        Discard
-      </Button>
-      <Button
-        variant="default"
-        size="xs"
-        onClick={() => {
-          actions.applyChanges();
-          onClose();
-        }}
-        disabled={!canApply}
-        title="Apply changes"
-      >
-        <Check className="mr-1 h-3 w-3" />
-        Apply
-      </Button>
-    </div>
-  );
-}
-
-/**
  * Renders a chart tool call with visualization using Vega-Lite.
  * Supports inline editing with AI slice persistence.
  *
@@ -97,9 +47,6 @@ export function VegaChartToolResult({
   editable = true,
   editorMode = 'both',
 }: VegaChartToolResultProps) {
-  const editorPopover = useDisclosure();
-  const [activeTab, setActiveTab] = useState<'spec' | 'sql'>('spec');
-
   // AI slice integration for persisting changes
   const setToolAdditionalData = useStoreWithAi(
     (s) => s.ai.setSessionToolAdditionalData,
@@ -138,10 +85,6 @@ export function VegaChartToolResult({
     [toolCallId, currentSessionId, appliedSpec, setToolAdditionalData],
   );
 
-  const showSpecEditor = editorMode === 'spec' || editorMode === 'both';
-  const showSqlEditor = editorMode === 'sql' || editorMode === 'both';
-  const showTabs = showSpecEditor && showSqlEditor;
-
   return (
     <div className={cn('flex flex-col gap-2', className)}>
       <VegaChartContainer
@@ -152,96 +95,14 @@ export function VegaChartToolResult({
         onSpecChange={handleSpecChange}
         onSqlChange={handleSqlChange}
       >
-        {/* Chart with edit popover */}
+        {/* Chart with actions toolbar */}
         <div className="relative min-h-[300px]">
-          <VegaChartDisplay aspectRatio={16 / 9}>
-            <VegaLiteArrowChart.Actions className="right-3 top-[-1px]">
+          <VegaChartDisplay aspectRatio={16 / 9} className="pt-2">
+            <VegaLiteArrowChart.Actions className="right-3">
               <VegaExportAction />
+              {editable && <VegaEditAction editorMode={editorMode} />}
             </VegaLiteArrowChart.Actions>
           </VegaChartDisplay>
-
-          {/* Edit button with popover */}
-          {editable && (
-            <Popover
-              open={editorPopover.isOpen}
-              onOpenChange={editorPopover.onToggle}
-            >
-              <PopoverTrigger asChild>
-                <Button
-                  className="absolute right-[40px] top-0"
-                  type="button"
-                  variant="ghost"
-                  size="xs"
-                  aria-label="Edit chart specification"
-                >
-                  <EditIcon className="h-4 w-4" />
-                </Button>
-              </PopoverTrigger>
-              <PopoverContent
-                align="end"
-                side="bottom"
-                className="w-[400px] p-0"
-              >
-                <div className="flex h-[400px] flex-col">
-                  {/* Header with tabs */}
-                  <div className="flex items-center justify-between border-b px-2 py-1">
-                    {showTabs ? (
-                      <Tabs
-                        value={activeTab}
-                        onValueChange={(v) => setActiveTab(v as 'spec' | 'sql')}
-                        className="w-full"
-                      >
-                        <div className="flex items-center justify-between">
-                          <TabsList className="h-7">
-                            <TabsTrigger
-                              value="spec"
-                              className="h-6 px-2 text-xs"
-                            >
-                              <Code2 className="mr-1 h-3 w-3" />
-                              Spec
-                            </TabsTrigger>
-                            <TabsTrigger
-                              value="sql"
-                              className="h-6 px-2 text-xs"
-                            >
-                              <Database className="mr-1 h-3 w-3" />
-                              SQL
-                            </TabsTrigger>
-                          </TabsList>
-                          <EditorActions onClose={editorPopover.onClose} />
-                        </div>
-                      </Tabs>
-                    ) : (
-                      <div className="flex w-full items-center justify-between">
-                        <span className="text-sm font-medium">
-                          {showSpecEditor ? 'Vega-Lite Spec' : 'SQL Query'}
-                        </span>
-                        <EditorActions onClose={editorPopover.onClose} />
-                      </div>
-                    )}
-                  </div>
-
-                  {/* Editor content */}
-                  <div className="flex-1 overflow-hidden">
-                    {showTabs ? (
-                      <Tabs value={activeTab} className="h-full">
-                        <TabsContent value="spec" className="mt-0 h-full">
-                          <VegaSpecEditorPanel title="" className="h-full" />
-                        </TabsContent>
-                        <TabsContent value="sql" className="mt-0 h-full">
-                          <VegaSqlEditorPanel title="" className="h-full" />
-                        </TabsContent>
-                      </Tabs>
-                    ) : showSpecEditor ? (
-                      <VegaSpecEditorPanel title="" className="h-full" />
-                    ) : (
-                      <VegaSqlEditorPanel title="" className="h-full" />
-                    )}
-                  </div>
-                </div>
-              </PopoverContent>
-            </Popover>
-          )}
         </div>
       </VegaChartContainer>
     </div>
