+++ packages/vega/src/index.ts
@@ -89,5 +89,7 @@ export type {VegaChartContextValue} from './VegaChartContext';
 // Export action components
 export {VegaExportAction} from './VegaExportAction';
 export type {VegaExportActionProps} from './VegaExportAction';
+export {VegaEditAction} from './VegaEditAction';
+export type {VegaEditActionProps} from './VegaEditAction';
 export {VegaChartActions} from './VegaChartActions';
 export type {VegaChartActionsProps} from './VegaChartActions';
